Wrote a performance test generation utility
Not yet finished, does not yet generate all necessary data for tests.

It would be a good idea to store this data in
SQLite files and then import those to MySQL if
necessary when running performance tests.

Idea is to allow generating multiple
different scenarios and then measure the
performance of different endpoints to estimate
time complexity and find bottlenecks.